Remove getServerSideProps on index page now uses-> getStaticProps

+++ components/Card.tsx
@@ -4,6 +4,9 @@ import { HeartIcon } from '@heroicons/react/solid'
 import { IHome } from '../types/home'
 import { useEffect, useState } from 'react'
 import { useDebouncedCallback } from 'use-debounce'
+import { useSWRConfig } from 'swr'
+import { useSession } from 'next-auth/react'
+
 interface ICardsProps extends IHome {
   isFavorite?: boolean
 }
@@ -19,6 +22,8 @@ const Card = ({
   isFavorite = false,
 }: ICardsProps) => {
   const [favorite, setFavorite] = useState(false)
+  const { data: session } = useSession()
+  const { mutate } = useSWRConfig()
 
   useEffect(() => {
     if (isFavorite) {
@@ -27,10 +32,14 @@ const Card = ({
   }, [isFavorite])
 
   const debouncedSubmitLike = useDebouncedCallback((homeID: string) => {
-    if (favorite) {
-      fetch(`/api/homes/${homeID}/favorite`, { method: 'PUT' })
-    } else {
-      fetch(`/api/homes/${homeID}/favorite`, { method: 'DELETE' })
+    if (session) {
+      if (favorite) {
+        fetch(`/api/homes/${homeID}/favorite`, { method: 'PUT' })
+      } else {
+        fetch(`/api/homes/${homeID}/favorite`, { method: 'DELETE' })
+      }
+      mutate(`/api/users/${session.user.id}/favorites`)
+      console.log("mutating")
     }
   }, 700)
   return (
